refactor: redefine `IPermission['_id']` as a string literal union type

+++ apps/meteor/server/startup/migrations/v305.ts
@@ -1,3 +1,4 @@
+import type { IPermission } from '@rocket.chat/core-typings';
 import { Permissions } from '@rocket.chat/models';
 
 import { addMigration } from '../../lib/migrations';
@@ -6,6 +7,6 @@ addMigration({
 	version: 305,
 	name: 'Remove unused view-history permission',
 	async up() {
-		await Permissions.deleteOne({ _id: 'view-history' });
+		await Permissions.deleteOne({ _id: 'view-history' as IPermission['_id'] });
 	},
 });
