+++ apps/meteor/server/startup/migrations/v331.ts
@@ -1,3 +1,4 @@
+import type { IPermission } from '@rocket.chat/core-typings';
 import { Permissions } from '@rocket.chat/models';
 
 import { addMigration } from '../../lib/migrations';
@@ -6,6 +7,6 @@ addMigration({
 	version: 331,
 	name: 'Remove view-livechat-current-chats permission',
 	async up() {
-		await Permissions.deleteOne({ _id: 'view-livechat-current-chats' });
+		await Permissions.deleteOne({ _id: 'view-livechat-current-chats' as IPermission['_id'] });
 	},
 });
